services/transactions-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-dotenv
passlib[bcrypt] # For JWT deps
python-jose[cryptography] # For JWT deps
pydantic-settings>=2.0.0
PyJWT>=2.8.0
python-multipart

# ML & Data
scikit-learn
pandas
joblib

# CSV
# (Handled by standard lib)